+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.30] - 2025-05-XX
+
+### Changed
+
+- Update PSScriptAnalyzerSettings.psd1 to exclude 'PSAvoidUsingWriteHost' rule
+- Refactor New-VeeamDiagram function to replace Write-ColorOutput with Write-Host for improved output handling
+- Update `Diagrammer.Core` module version to 0.2.27
+
+### Fixed
+
+- Fix Veeam Logo aspect ratio
+
 ## [0.6.29] - 2025-05-17
 
 ### Changed